Fix build on Ubuntu 16.04
The libcxxabi header files are packaged in '/usr/include/libcxxabi',
while the system include path is '/usr/include'. This would work if
libc++'s headers prefixed the libcxxabi directory name when they
included files from it, but they do not. This is Debian Bug #800492,
and it seems to have been inherited by Ubuntu 16.04.

Static linking against libgcc and libstdc++ in release mode is also
broken on Ubuntu 16.04 (at least in some configurations). There may
be a way to actually fix the problem, but for now the static linking
has been made an option that's enabled by default. If you encounter
linker errors in release mode, maybe disable PORTABLE_RELEASE.